refactor(runtime): split handle.rs into cohesive modules (§7-8)
`handle.rs` was 1670 lines, well over the 500-line module limit. Split it
along its natural seams into five files, each under 500 lines, with no
behavior change (the 55 fn signatures are identical before/after):

- `handle.rs` (301) — `impl IoHandleState` data-IO ops: buffered/raw file
  writes (`write_file_payload`, `write_all_to_file`, `native_write_bytes_file`),
  the VM-native read/write helpers (`read_line_native`, `slurp_string_native`,
  `read_bytes_native`, `read_chars_native`), and flush/target classification.
- `handle_io.rs` (290) — `impl Interpreter` output side: `write_to_handle_value*`,
  `write_bytes_to_handle_value`, `close_handle_value`, the `with_handle_mut*`
  confined-borrow helpers, `handle_id_from_value`, and line-separator parsing.
- `handle_read.rs` (357) — `impl Interpreter` record/line/word/byte reading
  (`read_record_*`, `read_line_from_handle_value`, `read_word_from_handle_value`,
  `read_bytes_from_handle_value`).
- `handle_read_chars.rs` (249) — `impl Interpreter` character reading and
  UTF-8/UTF-16 decode (`read_utf8_char`, `read_utf16_char`,
  `read_chars_from_handle_value`).
- `handle_open.rs` (491) — `impl IoHandleState` positioning/settings
  (`tell`/`eof`/`seek`/`close`, chomp/nl-out/out-buffer/encoding/descriptor)
  plus `impl Interpreter` positioning wrappers, open-flag parsing
  (`parse_io_flags_values`), and `open_file_handle`.

`normalize_line_separators` and `parse_nl_in_value` move from private `fn` to
`pub(super)` (now called across the handle_io/handle_open boundary). The
`num_traits::ToPrimitive` import moves to `handle_io.rs` (its only user,
`parse_out_buffer_size`). Fourth slice of the §7-8 file-splitting sweep.
