+++ src/runtime/mod.rs
@@ -153,6 +153,10 @@ pub(crate) mod did_you_mean;
 mod dispatch;
 mod eval_check;
 mod handle;
+mod handle_io;
+mod handle_open;
+mod handle_read;
+mod handle_read_chars;
 mod io;
 mod io_handles;
 pub(crate) mod json;
